+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+### 0.2.0
+
+- Port to Dart null safety.
+- Add String extension methods `toUpperCaseTr`, `toLowerCaseTr`, `toTitleCaseTr`
+
 ### 0.1.0+2
 
-- Remove new keyword, use isEmpty and isNotEmpty, small fixes.
+- Remove the `new` keyword, use `isEmpty` and `isNotEmpty`, small fixes.
 - Update dependencies.
 
 ### 0.1.0+1